ReaScript function export: harden SNM_GetProjectMarkerName()
i.e. validate/keep track of created WDL_FastString instances

Also move some ReaScript-only functions to SnM_Misc.cpp
