+++ ReaScript.cpp
@@ -27,7 +27,6 @@
 
 #include "stdafx.h"
 #include "SnM/SnM_FX.h"
-#include "SnM/SnM_Marker.h"
 #include "SnM/SnM_Misc.h"
 #include "SnM/SnM_Resources.h"
 #include "SnM/SnM_Routing.h"
